Started implementing my sets page
