--- src/app/Auth/Login/login.component.css
@@ -1,24 +1,0 @@
-.fields {
-  display: flex;
-  flex-direction: column;
-  row-gap: 12px;
-  margin-top: 32px;
-}
-.buttons {
-  display: flex;
-  flex-direction: column;
-  row-gap: 12px;
-  margin-top: 32px;
-}
-
-.container {
-  width: 450px
-}
-
-.forgot-link{
-  text-decoration: none;
-}
-
-.forgot-link:visited{
-  color: blueviolet;
-}
