+++ src/app/Global/Auth/Login/login.component.css
@@ -1,0 +1,16 @@
+.fields {
+  display: flex;
+  flex-direction: column;
+  row-gap: 12px;
+  margin-top: 32px;
+}
+.buttons {
+  display: flex;
+  flex-direction: column;
+  row-gap: 12px;
+  margin-top: 32px;
+}
+
+.container {
+  width: 450px
+}
