+++ src/app/Global/Auth/Login/login.component.html
@@ -1,0 +1,26 @@
+<div class="container">
+  <h1 class="mat-typography">Hey there!</h1>
+  <h3 class="mat-typography">We need you to log in before using 勉強！</h3>
+  <div class="fields">
+    <mat-form-field appearance="outline">
+      <mat-label>Email</mat-label>
+      <input matInput placeholder="youremail@mail.com" [formControl]="loginControl">
+      <mat-error *ngIf="loginControl.invalid">{{getEmailErrorMessage()}}</mat-error>
+    </mat-form-field>
+    <mat-form-field appearance="outline">
+      <mat-label>Password</mat-label>
+      <input matInput placeholder="Your very secure password" [formControl]="passwordControl"
+             [type]="isPasswordHidden ? 'password' : 'text'">
+      <button mat-icon-button matSuffix (click)="isPasswordHidden=!isPasswordHidden">
+        <mat-icon fontSet="material-icons-outlined">{{isPasswordHidden ? "visibility_off" : "visibility_on"}}</mat-icon>
+      </button>
+      <mat-error *ngIf="passwordControl.invalid">{{getPasswordErrorMessage()}}</mat-error>
+    </mat-form-field>
+    <a routerLink="/auth/forgot-password">I forgot my password</a>
+  </div>
+  <div class="buttons">
+    <button mat-flat-button color="primary" (click)="onLoginClicked()">Login</button>
+    <p class="mat-typography">Don't have an account yet?</p>
+    <button mat-flat-button color="accent" (click)="onRegistrationClicked()">Sign up</button>
+  </div>
+</div>
