+++ src/app/Global/Auth/Registration/registration.component.css
