+++ src/app/Global/Hub/Components/Set/set.component.css
@@ -1,0 +1,22 @@
+set {
+max-width: 360px;
+max-height: 480px;
+padding: 16px;
+position: relative;
+width: 360px;
+height: 480px;
+border-radius: 12px;
+border-color: #81898a;
+border-width: 1px;
+border-style: solid;
+display: flex;
+flex-direction: column;
+justify-content: space-between;
+transition: all 0.3s cubic-bezier(.25,.8,.25,1);
+box-sizing: border-box;
+}
+
+set:hover{
+  box-shadow: 0 14px 28px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22);
+  cursor: inherit;
+}
