+++ src/app/Global/Hub/Pages/Account/account.component.html
@@ -1,0 +1,4 @@
+<h1 class="mat-display-1" style="margin: 0">Account</h1>
+<div class="page-content">
+
+</div>
