+++ src/app/Global/Hub/Pages/AllSets/all-sets.component.html
@@ -1,0 +1,4 @@
+<h1 class="mat-display-1" style="margin: 0">All Sets</h1>
+<div class="page-content">
+
+</div>
