+++ src/app/Global/Hub/Pages/Home/home-page.component.html
@@ -1,0 +1,9 @@
+<h1 class="mat-display-1" style="margin: 0">Welcome to the 勉強!</h1>
+<div class="page-content">
+  <p>Easy and convenient service to learn kanji! Create your own sets, share it with friends, search for
+    pre-made sets or public sets of other users</p>
+  <p>TODO i guess</p>
+  <p>Go to <a routerLink="/hub/my-sets">My Sets</a> page</p>
+  <p>Create new set</p>
+  <p>Study!</p>
+</div>
