+++ src/app/Global/Hub/Pages/Home/home-page.component.ts
@@ -1,0 +1,9 @@
+import {Component, EventEmitter, Output} from "@angular/core";
+
+@Component({
+  selector: "home-page",
+  templateUrl: "home-page.component.html",
+  styleUrls: ["home-page.component.css"]
+})
+
+export class HomePageComponent {}
