+++ src/app/Global/Hub/Pages/MySets/my-sets.component.css
@@ -1,0 +1,5 @@
+.top-section{
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
