+++ src/app/Global/Hub/Pages/MySets/my-sets.component.html
@@ -1,0 +1,7 @@
+<div class="top-section">
+  <h1 class="mat-display-1" style="margin: 0">My Sets</h1>
+  <button mat-raised-button color="primary" (click)="onCreateNewSetClicked()">Create new set</button>
+</div>
+<div class="page-content">
+
+</div>
