+++ src/app/Global/Hub/hub.component.css
@@ -1,0 +1,34 @@
+mat-sidenav-container{
+  height: 100%;
+}
+
+mat-sidenav-content{
+  height: 100%;
+}
+
+mat-sidenav{
+  width: 260px;
+  border-bottom-right-radius: 12px;
+  border-top-right-radius: 12px;
+}
+
+.text-padding{
+  padding-left: 8px;
+}
+
+.no-margin{
+  margin: 0;
+}
+
+.pageContainer{
+  height: 100%;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  padding: 45px;
+}
+
+mat-nav-list{
+  margin-left: 8px;
+  margin-right: 8px;
+}
