+++ src/app/Global/NotFound/not-found.component.html
@@ -1,0 +1,1 @@
+<p>not found</p>
