+++ src/app/Guards/auth.guard.ts
@@ -1,0 +1,12 @@
+import {ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot, UrlTree} from "@angular/router";
+import {Observable} from "rxjs";
+
+export class AuthGuard implements CanActivate{
+  constructor() {
+    //INJECT HERE
+  }
+  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    //TODO inject redux and check if user is authorised
+    return true;
+  }
+}
