--- src/app/Hub/Pages/Account/account.component.html
@@ -1,1 +1,0 @@
-<p>hello world account</p>
