--- src/app/Hub/Pages/AllSets/all-sets.component.html
@@ -1,1 +1,0 @@
-<p>hello world all sets</p>
