--- src/app/Hub/Pages/Home/home-page.component.html
@@ -1,1 +1,0 @@
-<p>Hello home world!</p>
