--- src/app/Hub/Pages/Home/home-page.component.ts
@@ -1,9 +1,0 @@
-import {Component} from "@angular/core";
-
-@Component({
-  selector: "home-page",
-  templateUrl: "home-page.component.html",
-  styleUrls: ["home-page.component.css"]
-})
-
-export class HomePageComponent {}
