--- src/app/Hub/Pages/MySets/my-sets.component.html
@@ -1,1 +1,0 @@
-<p>hello world my sets</p>
