--- src/app/Hub/Pages/MySets/my-sets.component.ts
@@ -1,11 +1,0 @@
-import {Component} from "@angular/core";
-
-@Component({
-  selector: "my-sets-page",
-  templateUrl: "my-sets.component.html",
-  styleUrls: ["my-sets.component.css"]
-})
-
-export class MySetsComponent {
-
-}
