--- src/app/Hub/hub.component.css
@@ -1,17 +1,0 @@
-mat-sidenav-container{
-  height: 100%;
-}
-
-mat-sidenav{
-  width: 240px;
-  border-bottom-right-radius: 12px;
-  border-top-right-radius: 12px;
-}
-
-.text-padding{
-  padding-left: 8px;
-}
-
-.no-margin{
-  margin: 0;
-}
