--- src/app/Hub/hub.component.ts
@@ -1,9 +1,0 @@
-import {Component} from "@angular/core";
-
-@Component({
-  selector: "hub-component",
-  styleUrls: ["hub.component.css"],
-  templateUrl: "hub.component.html"
-})
-
-export class HubComponent {}
