+++ src/app/Models/Kanji.ts
@@ -1,0 +1,11 @@
+export default class Kanji {
+  kanji: string;
+  kunyomi: string[]
+  onyomi: string[]
+
+  constructor(kanji: string, kunyomi: string[], onyomi: string[]) {
+    this.kanji = kanji
+    this.kunyomi = kunyomi
+    this.onyomi = onyomi
+  }
+}
