+++ src/styles.css
@@ -1,17 +1,39 @@
 /* You can add global styles to this file, and also import other style files */
 
-html, body { height: 100%; }
-body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
+html, body {
+  height: 100%;
+}
 
-.mat-horizontal-stepper-wrapper{
+body {
+  margin: 0;
+  font-family: Roboto, "Helvetica Neue", sans-serif;
+}
+
+.mat-horizontal-stepper-wrapper {
   row-gap: 16px;
 }
 
-mat-list-item{
+mat-list-item {
   border-radius: 14px;
 }
 
-.mat-list-item-ripple{
+.route-selected {
+  background: rgba(255, 255, 255, .04);
+}
+
+.mat-list-item-ripple {
   border-radius: 14px;
+}
+
+a{
+  text-decoration: none;
+}
+
+a:visited{
+  color: blueviolet;
+}
+
+.page-content{
+  margin-top: 24px;
 }
